pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "OBRAG"
authors = [
    { name = "Angel Patricio", email = "aopatric@mit.edu" }
]
description = "Obsidian RAG assistant with access to your local vault."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "tomli-w>=1.2.0, <2.0",
    "tqdm>=4.67.1, <5.0",
    "langchain-chroma>=0.2.4, <0.3",
    "langchain-openai>=0.3.24, <0.4",
    "langchain-huggingface>=0.3.0, <0.4",
    "langchain-community>=0.3.26, <0.4",
    "langgraph>=0.4.8, <0.5",
    "sentence-transformers>=4.1.0, <4.2"
]
version = "0.1.0"

[project.scripts]
obrag = "obrag.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["obrag*"]